requirements.txt
# Health Advisor Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Anthropic Claude API
anthropic==0.39.0

# MCP SDK
mcp==1.2.0

# Database
aiosqlite==0.20.0

# Configuration
pyyaml==6.0.1

# HTTP Client (for MCP SSE transport)
httpx==0.27.2
httpx-sse==0.4.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1
